- Add duty_payment and duty_payment_account to shipment options - Move shipment options to its own file
